Progress on implementing extra link in the plugin overview.
