+++ admin/class-admin.php
@@ -212,6 +212,8 @@ class WPSEO_Admin {
 	 * @return array
 	 */
 	public function add_action_link( $links, $file ) {
+		$first_time_configuration_notice_helper = \YoastSEO()->helpers->first_time_configuration_notice;
+
 		if ( $file === WPSEO_BASENAME && WPSEO_Capability_Utils::current_user_can( 'wpseo_manage_options' ) ) {
 			if ( is_network_admin() ) {
 				$settings_url = network_admin_url( 'admin.php?page=' . self::PAGE_IDENTIFIER );
@@ -227,6 +229,15 @@ class WPSEO_Admin {
 		$faq_link = '<a href="' . esc_url( WPSEO_Shortlinker::get( 'https://yoa.st/1yc' ) ) . '" target="_blank">' . __( 'FAQ', 'wordpress-seo' ) . '</a>';
 		array_unshift( $links, $faq_link );
 
+		if ( $first_time_configuration_notice_helper->first_time_configuration_not_finished() ) {
+			/* translators: CTA to finish the first time configuration. %s: Network title. */
+			$message = __( sprintf( 'Finish your %s', strtolower($first_time_configuration_notice_helper->get_first_time_configuration_title() )), 'wordpress-seo' );
+			$ftc_link = '<a href="' . esc_url( admin_url( 'admin.php?page=wpseo_dashboard#top#first-time-configuration' ) ) . '" target="_blank">' . $message . '</a>';
+			array_unshift( $links, $ftc_link );
+		}
+
+
+
 		$addon_manager = new WPSEO_Addon_Manager();
 		if ( YoastSEO()->helpers->product->is_premium() ) {
 
@@ -234,9 +245,9 @@ class WPSEO_Admin {
 			unset( $links['deactivate'] );
 			$no_deactivation_explanation = '<span style="color: #32373c">' . sprintf(
 				/* translators: %s expands to Yoast SEO Premium. */
-				__( 'Required by %s', 'wordpress-seo' ),
-				'Yoast SEO Premium'
-			) . '</span>';
+					__( 'Required by %s', 'wordpress-seo' ),
+					'Yoast SEO Premium'
+				) . '</span>';
 
 			array_unshift( $links, $no_deactivation_explanation );
 
@@ -331,7 +342,7 @@ class WPSEO_Admin {
 			[
 				'isRtl'                   => is_rtl(),
 				'variable_warning'        => sprintf(
-					/* translators: %1$s: '%%term_title%%' variable used in titles and meta's template that's not compatible with the given template, %2$s: expands to 'HelpScout beacon' */
+				/* translators: %1$s: '%%term_title%%' variable used in titles and meta's template that's not compatible with the given template, %2$s: expands to 'HelpScout beacon' */
 					__( 'Warning: the variable %1$s cannot be used in this template. See the %2$s for more info.', 'wordpress-seo' ),
 					'<code>%s</code>',
 					'HelpScout beacon'
