+++ src/integrations/admin/first-time-configuration-notice-integration.php
@@ -8,6 +8,7 @@ use Yoast\WP\SEO\Helpers\Indexing_Helper;
 use Yoast\WP\SEO\Helpers\Options_Helper;
 use Yoast\WP\SEO\Integrations\Integration_Interface;
 use Yoast\WP\SEO\Presenters\Admin\Notice_Presenter;
+use Yoast\WP\SEO\Helpers\First_Time_Configuration_Notice_Helper;
 
 /**
  * First_Time_Configuration_Notice_Integration class
@@ -22,13 +23,6 @@ class First_Time_Configuration_Notice_Integration implements Integration_Interfa
 	private $options_helper;
 
 	/**
-	 * The indexing helper.
-	 *
-	 * @var Indexing_Helper
-	 */
-	private $indexing_helper;
-
-	/**
 	 * The admin asset manager.
 	 *
 	 * @var WPSEO_Admin_Asset_Manager
@@ -36,11 +30,9 @@ class First_Time_Configuration_Notice_Integration implements Integration_Interfa
 	private $admin_asset_manager;
 
 	/**
-	 * Whether we show the alternate mesage.
-	 *
-	 * @var bool
+	 * @var \Yoast\WP\SEO\Helpers\First_Time_Configuration_Notice_Helper
 	 */
-	private $show_alternate_message;
+	private $first_time_configuration_notice_helper;
 
 	/**
 	 * {@inheritDoc}
@@ -53,18 +45,16 @@ class First_Time_Configuration_Notice_Integration implements Integration_Interfa
 	 * First_Time_Configuration_Notice_Integration constructor.
 	 *
 	 * @param Options_Helper            $options_helper      The options helper.
-	 * @param Indexing_Helper           $indexing_helper     The indexing helper.
 	 * @param WPSEO_Admin_Asset_Manager $admin_asset_manager The admin asset manager.
 	 */
 	public function __construct(
 		Options_Helper $options_helper,
-		Indexing_Helper $indexing_helper,
+		First_Time_Configuration_Notice_Helper $first_time_configuration_notice_helper,
 		WPSEO_Admin_Asset_Manager $admin_asset_manager
 	) {
-		$this->options_helper         = $options_helper;
-		$this->indexing_helper        = $indexing_helper;
-		$this->admin_asset_manager    = $admin_asset_manager;
-		$this->show_alternate_message = false;
+		$this->options_helper                         = $options_helper;
+		$this->admin_asset_manager                    = $admin_asset_manager;
+		$this->first_time_configuration_notice_helper = $first_time_configuration_notice_helper;
 	}
 
 	/**
@@ -90,37 +80,7 @@ class First_Time_Configuration_Notice_Integration implements Integration_Interfa
 	 * @return bool Whether the "First-time SEO Configuration" admin notice should be displayed.
 	 */
 	public function should_display_first_time_configuration_notice() {
-		if ( ! $this->options_helper->get( 'dismiss_configuration_workout_notice', false ) === false ) {
-			return false;
-		}
-
-		if ( ! $this->user_can_do_first_time_configuration() ) {
-			return false;
-		}
-
-		if ( ! $this->on_wpseo_admin_page_or_dashboard() ) {
-			return false;
-		}
-
-		if ( $this->is_first_time_configuration_finished() ) {
-			return false;
-		}
-
-		if ( $this->options_helper->get( 'first_time_install', false ) !== false ) {
-			return ! $this->are_site_representation_name_and_logo_set() || $this->indexing_helper->get_unindexed_count() > 0;
-		}
-
-		if ( $this->indexing_helper->is_initial_indexing() === false ) {
-			return false;
-		}
-
-		if ( $this->indexing_helper->is_finished_indexables_indexing() === true ) {
-			return false;
-		}
-
-		$this->show_alternate_message = true;
-
-		return ! $this->are_site_representation_name_and_logo_set();
+		return $this->first_time_configuration_notice_helper->should_display_first_time_configuration_notice();
 	}
 
 	/**
@@ -135,8 +95,8 @@ class First_Time_Configuration_Notice_Integration implements Integration_Interfa
 
 		$this->admin_asset_manager->enqueue_style( 'monorepo' );
 
-		$title = ( ! $this->show_alternate_message ) ? \__( 'First-time SEO configuration', 'wordpress-seo' ) : \__( 'SEO configuration', 'wordpress-seo' );
-		if ( ! $this->show_alternate_message ) {
+		$title = $this->first_time_configuration_notice_helper->get_first_time_configuration_title();
+		if ( ! $this->first_time_configuration_notice_helper->should_show_alternate_message() ) {
 			$content = \sprintf(
 				/* translators: 1: Link start tag to the first-time configuration, 2: Yoast SEO, 3: Link closing tag. */
 				\__( 'Get started quickly with the %1$s%2$s First-time configuration%3$s and configure Yoast SEO with the optimal SEO settings for your site!', 'wordpress-seo' ),
@@ -184,76 +144,5 @@ class First_Time_Configuration_Notice_Integration implements Integration_Interfa
 				} );
 			} );
 			</script>";
-	}
-
-	/**
-	 * Whether the user can do the first-time configuration.
-	 *
-	 * @return bool Whether the current user can do the first-time configuration.
-	 */
-	private function user_can_do_first_time_configuration() {
-		return \current_user_can( 'wpseo_manage_options' );
-	}
-
-	/**
-	 * Whether the user is currently visiting one of our admin pages or the WordPress dashboard.
-	 *
-	 * @return bool Whether the current page is a Yoast SEO admin page
-	 */
-	private function on_wpseo_admin_page_or_dashboard() {
-		$pagenow = $GLOBALS['pagenow'];
-
-		// Show on the WP Dashboard.
-		if ( $pagenow === 'index.php' ) {
-			return true;
-		}
-
-		$page_from_get = \filter_input( \INPUT_GET, 'page' );
-
-		// Show on Yoast SEO pages, with some exceptions.
-		if ( $pagenow === 'admin.php' && \strpos( $page_from_get, 'wpseo' ) === 0 ) {
-			$exceptions = [
-				'wpseo_installation_successful',
-				'wpseo_installation_successful_free',
-			];
-
-			if ( ! \in_array( $page_from_get, $exceptions, true ) ) {
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	/**
-	 * Whether all steps of the first-time configuration have been finished.
-	 *
-	 * @return bool Whether the first-time configuration has been finished.
-	 */
-	private function is_first_time_configuration_finished() {
-		$configuration_finished_steps = $this->options_helper->get( 'configuration_finished_steps', [] );
-
-		return \count( $configuration_finished_steps ) === 3;
-	}
-
-	/**
-	 * Whether the site representation name and logo have been set.
-	 *
-	 * @return bool  Whether the site representation name and logo have been set.
-	 */
-	private function are_site_representation_name_and_logo_set() {
-		$company_or_person = $this->options_helper->get( 'company_or_person', '' );
-
-		if ( $company_or_person === '' ) {
-			return false;
-		}
-
-		if ( $company_or_person === 'company' ) {
-			return ! empty( $this->options_helper->get( 'company_name' ) )
-					&& ! empty( $this->options_helper->get( 'company_logo', '' ) );
-		}
-
-		return ! empty( $this->options_helper->get( 'company_or_person_user_id' ) )
-				&& ! empty( $this->options_helper->get( 'person_logo', '' ) );
 	}
 }
